Top Level Object should not have a container
Data::Walk Version 2.01 assigns `container` and `type` for top-level objects
in an inconsistent way. On one hand, A scalar at top level gets the
arrayref of `@args` as `container`, and `ARRAY` as `type`. Whereas
a hashref, or arrayref, at top-level gets *ITSELF* as the `container`,
and its own ref type (`ARRAY|HASH`) as `type`. The latter behaviour
is not consistent because it implies that the item is contained in
itself.

A set of rules for consistent behaviour regarding top-level objects can be:
- Top level object passed to `walk` has `$Data::Walk::depth == 1`.
- `$Data::Walk::container` and `$Data::Walk::type` is `undef` for such objects.
  (They have no container, and hence, no container type.)

A patch against commit b3f9af86587a513d9a3a1e8f318ef6a30e328872 is included
here.

Summary of Changes:

- new file:   t/05bug-container-type-at-depth-one.t

  Tests the above consistent behaviour for
  the following calls of `Data::Walk::walk`:
    - Single hashref argument
      `walk sub { ... }, $hashref;`
    - Single arrayref argument
      `walk sub { ... }, $arrayref;`
    - Many argument with the first one as scalar
      `walk sub { ... }, 5, $hash,$arrayref;`

- modified:   t/04bug-container-type-by-depth.t

  Makes the test compliant to the above mentioned consistent behaviour.

- modified: lib/Data/Walk.pm

  Fix setting of Top-Level Objects' container